% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_dataset.R
\name{choose_dataset}
\alias{choose_dataset}
\title{#########################################################################################
Function to choose a project dataset}
\usage{
choose_dataset(caption = "Select a dataset")
}
\arguments{
\item{caption}{a character string}
}
\value{
dir
}
\description{
Patrice Vallin, easyXgboost, Apr 2020.
########################################################################################
}
