% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colScale.R
\name{colScale}
\alias{colScale}
\title{#########################################################################################
scale dataset}
\usage{
colScale(
  x,
  center = TRUE,
  scale = TRUE,
  add_attr = TRUE,
  rows = NULL,
  cols = NULL
)
}
\arguments{
\item{dataFile}{a character string specifying the location of the project dataset}
}
\description{
Adapted from : https://www.r-bloggers.com/a-faster-scale-function/
########################################################################################
}
