% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_initialize.R
\name{easyXgboost.initialize}
\alias{easyXgboost.initialize}
\title{#########################################################################################
Initialize easyXgboost}
\usage{
easyXgboost.initialize()
}
\description{
Patrice Vallin, easyXgboost, Apr 2020.
########################################################################################
}
