% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_helper.R
\name{findElbow}
\alias{findElbow}
\title{SSE and findElbow function are adapted from FlowSOM R package
Sofie Van Gassen et al., Cytometry Part A, 2015
https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.22625
http://bioconductor.org/packages/release/bioc/html/FlowSOM.html}
\usage{
findElbow(data)
}
\arguments{
\item{data}{a array of numeric values}
}
\value{
optimal
}
\description{
SSE and findElbow function are adapted from FlowSOM R package
Sofie Van Gassen et al., Cytometry Part A, 2015
https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.22625
http://bioconductor.org/packages/release/bioc/html/FlowSOM.html
}
