% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_parameters.R
\name{import_parameters}
\alias{import_parameters}
\title{Load & save analysis parameters}
\usage{
import_parameters(dataFile)
}
\arguments{
\item{dataFile}{a character string specifying the location of the project dataset}
}
\description{
Load & save analysis parameters
}
