% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_helper.R
\name{prediction_model}
\alias{prediction_model}
\title{Construct a xgboost model based on cross validation}
\usage{
prediction_model(
  dataset_stim,
  outcome_var,
  ErrorsHyperparameters,
  classif_type,
  seed,
  k
)
}
\arguments{
\item{dataset_stim}{a matrix}

\item{outcome_var}{an array}

\item{ErrorsHyperparameters}{a matrix}

\item{classif_type}{a character string}

\item{seed}{a numeric value}

\item{k}{a numeric value}
}
\value{
xgb
}
\description{
Patrice Vallin, easyXgboost, Apr 2020
}
