% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare the dataset}
\usage{
prepare_dataset(dataFile)
}
\arguments{
\item{dataFile}{a character string specifying the location of the project dataset}
}
\description{
Prepare the dataset
}
