% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_analysis2.R
\name{xgboost_analysis2}
\alias{xgboost_analysis2}
\title{Analyse the dataset with xgboost}
\usage{
xgboost_analysis2(dataFile)
}
\arguments{
\item{dataFile}{a character string specifying the location of the project dataset}
}
\description{
Patrice Vallin, easyXgboost, Apr 2020
.
SSE and findElbow function are adapted from FlowSOM R package
Sofie Van Gassen et al., Cytometry Part A, 2015
https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.22625
http://bioconductor.org/packages/release/bioc/html/FlowSOM.html
}
