% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_helper.R
\name{xgboost_cv}
\alias{xgboost_cv}
\title{Evaluate a xgboost model through cross validation}
\usage{
xgboost_cv(data_DM, xgb_pars)
}
\arguments{
\item{data_DM}{an xgboost pointer}

\item{xgb_pars}{a matrix}
}
\value{
ErrorsHyperparameters
}
\description{
Patrice Vallin, easyXgboost, Apr 2020
}
