% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_helper.R
\name{xgboost_hypertuning}
\alias{xgboost_hypertuning}
\title{Find optimal xgboost parameters through hypertuning}
\usage{
xgboost_hypertuning(data_DM, xgb_pars)
}
\arguments{
\item{data_DM}{an xgboost pointer}

\item{xgb_pars}{a matrix}
}
\description{
Patrice Vallin, easyXgboost, Apr 2020
}
